fix(models): finish freeing the guards from the CLI extras
Second review round on #184.

The previous round moved the Munk and Stommel guards into
``somax._src.core.resolution`` but left every ``from_nondimensional``
importing them from ``somax._src.cli._assertions``, which imports
``loguru`` at module level — declared only in the optional ``sim``
group. So the base-install failure the move was meant to fix was still
there, and ``check_deformation_radius`` was still defined in the CLI
module besides. It moves too, all three factories import from the base
package, and ``_assertions`` re-exports so the registry and existing
call sites are unchanged. A test runs each factory in a subprocess
with ``loguru`` forced to fail.

Also: ``if not burger`` raises on a multi-element NumPy or JAX array —
ambiguous truth value — so array-backed sequences were rejected before
the conversion that would have handled them. It tests ``len(...) == 0``
instead.
